fix(ci): split upload helpers + WorkerCtx to satisfy clippy + size gates
- run_worker now takes a WorkerCtx struct instead of 9 individual args
  (clippy's too_many_arguments fires at 7+).
- db/mod.rs was 1039 lines; split upload telemetry helpers + picker
  + list_recent_uploads into db/upload.rs.
- db/tests.rs was 1116 lines; moved the 5 upload tests into
  db/upload_tests.rs.

Closes nothing on its own — makes the #118 PR mergeable.
